Move cross_join_dfs from session.py into operations.py
Reviewer flagged that this single-loop wrapper around cross_join
belonged next to its primitive. The function is now alongside
cross_join in operations.py, and session.py no longer needs to import
cross_join.

Two import sites updated (config/project_config.py, test_ibis_session).
Completes the Phase 3 split of session.py: original 1457 lines now
sit at ~785, with the extracted concerns living in _duckdb_shims.py,
io.py, tz.py, null_checks.py, models.py, and the existing operations
/ types / temp modules.
